% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontawesome.R
\name{fontawesome}
\alias{fontawesome}
\title{fontawesome}
\usage{
fontawesome(aliases)
}
\arguments{
\item{aliases}{aliases}
}
\value{
text
}
\description{
convert fontawesome aliases to text
}
\author{
ygc
}

