% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_emoji.R
\name{geom_emoji}
\alias{geom_emoji}
\title{geom_emoji}
\usage{
geom_emoji(alias = "rose", color = "#67001f", size = 120, x = 1,
  y = 1, ...)
}
\arguments{
\item{alias}{emoji alias}

\item{color}{color}

\item{size}{size}

\item{x}{x position}

\item{y}{y position}

\item{...}{additional parameter for geom_text}
}
\value{
ggplot2 layer
}
\description{
wrapper layer for visualizing emoji
}
\author{
guangchuang yu
}

