% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_emoji.R
\name{geom_fontawesome}
\alias{geom_fontawesome}
\title{geom_fontawesome}
\usage{
geom_fontawesome(alias = "fa-github", color = "#67001f", size = 120,
  x = 1, y = 1, ...)
}
\arguments{
\item{alias}{fontawesome alias}

\item{color}{color}

\item{size}{size}

\item{x}{x position}

\item{y}{y position}

\item{...}{additional parameter for geom_text}
}
\value{
ggplot2 layer
}
\description{
wrapper layer for visualizing fontawesome
}
\author{
guangchuang yu
}

