% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emojifont.R
\name{load.emojifont}
\alias{load.emojifont}
\title{load.emojifont}
\usage{
load.emojifont(font = "EmojiOne.ttf")
}
\arguments{
\item{font}{one of output from list.emojifonts}
}
\description{
load emoji font
}
\author{
ygc
}
