% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emojifont.R
\name{sample_emoji}
\alias{sample_emoji}
\title{sample_emoji}
\usage{
sample_emoji(size, replace = FALSE)
}
\arguments{
\item{size}{a non-negative integer giving the number of items to choose}

\item{replace}{Should sampling be with replacement?}
}
\value{
random emoji
}
\description{
random emoji
}
\examples{
sample_emoji(3)
}
\author{
Guangchuang Yu
}
