% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontawesome.R
\name{sample_fontawesome}
\alias{sample_fontawesome}
\title{sample_fontawesome}
\usage{
sample_fontawesome(size, replace = FALSE)
}
\arguments{
\item{size}{a non-negative integer giving the number of items to choose}

\item{replace}{Should sampling be with replacement?}
}
\value{
random fontawesome
}
\description{
random fontawesome
}
\examples{
sample_fontawesome(3)
}
\author{
guangchuang yu
}

