% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emojifont.R
\name{search_emoji}
\alias{search_emoji}
\title{search_emoji}
\usage{
search_emoji(str, type = "aliases", approximate = FALSE)
}
\arguments{
\item{str}{string text}

\item{type}{one of 'aliases', 'description' and 'tags'}

\item{approximate}{logical}
}
\value{
corresponding aliases, can be used to get unicode by emoji function
}
\description{
search emoji
}
\author{
ygc
}
