% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontawesome.R
\name{search_fontawesome}
\alias{search_fontawesome}
\title{search_fontawesome}
\usage{
search_fontawesome(str, approximate = FALSE)
}
\arguments{
\item{str}{string text}

\item{approximate}{logical}
}
\value{
corresponding aliases
}
\description{
search fontawesome
}
\author{
ygc
}
