% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontawesome.R
\name{load.fontawesome}
\alias{load.fontawesome}
\title{load.fontawesome}
\usage{
load.fontawesome(font = "fontawesome-webfont.ttf")
}
\arguments{
\item{font}{font}
}
\description{
load fontawesome
}
\author{
ygc
}

